finished work

+++ src/main/kotlin/ar/edu/unq/desapp/grupoc/backenddesappapi/webservice/dto/CryptoStockDTO.kt
@@ -1,8 +1,10 @@
 package ar.edu.unq.desapp.grupoc.backenddesappapi.webservice.dto
 
+import java.math.BigDecimal
+
 data class CryptoStockDTO (
     val symbol: String,
-    val price: Double,
+    val price: BigDecimal,
     val quantity: Double,
-    val localPrice: Double
+    val localPrice: BigDecimal
 )
